requirements.txt
flask
joblib
numpy
scikit-learn
mysql-connector-python